Fixed double api call and migrated it to modal
